[rename] auth 패키지를 domain 패키지 아래로 리네이밍
