+++ src/main/java/Ness/Backend/domain/auth/security/AuthDetailService.java
@@ -1,6 +1,6 @@
 package Ness.Backend.domain.auth.security;
 
-import Ness.Backend.domain.entity.Member;
+import Ness.Backend.domain.member.entity.Member;
 import Ness.Backend.domain.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetailsService;
